requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
neo4j==5.27.0
google-generativeai==0.8.4
python-dotenv==1.0.1
pdfplumber==0.11.4
